// src/shared/interfaces/click-courses.interface.ts
export interface IClickDataCourses {
  id: number | null;
  Nombre_Corto: string | null;
  Nombre_Curso: string | null;
  Categoria: string | null;
  Categoria1: string | null;
  tipo_curso: string | null;
  Fecha: string | null;
  Anio: number | null;
  Mes: number | null;
  Dia: number | null;
  Usuarios: number | null;
  Clics: number | null;
}